requirements.txt
Pillow
transformers
torch
